finding friends frontend

+++ db/migrate/20130330053042_create_authorizations.rb
@@ -1,17 +1,26 @@
 class CreateAuthorizations < ActiveRecord::Migration
-  def change
+  def up
     create_table :authorizations do |t|
       t.string  :provider
       t.string  :uid
       t.integer :user_id
       t.string  :access_token
+      t.datetime  :expires_at
       t.text    :meta
 
       t.timestamps
     end
 
-    add_index :authorizations, :user_id
+    if ActiveRecord::Base.connection.tables.include?(:authentications)
+      drop_table :authentications
+    end
 
-    drop_table :authentications
+    add_index :authorizations, :user_id
+  end
+
+  def down
+    drop_table :authorizations
   end
 end
+
+
